+++ src/app/login-page/login-page.page.html
@@ -1,7 +1,7 @@
 <ion-content padding class="vertical-center">
   <img src="../../assets/barbell-outline.png" alt="" />
   <h1>DevFitness</h1>
-  <ion-button size="large">
+  <ion-button size="large" (click)="login()">
     <ion-icon slot="start" name="logo-google"></ion-icon>
     Entrar com Google</ion-button
   >
