+++ src/app/login-page/login-page.page.ts
@@ -1,5 +1,13 @@
 import { Component, OnInit } from '@angular/core';
-
+import {
+  getAuth,
+  Auth,
+  signInWithEmailAndPassword,
+  createUserWithEmailAndPassword,
+  signOut,
+  signInWithPopup,
+  GoogleAuthProvider,
+} from 'firebase/auth';
 @Component({
   selector: 'app-login-page',
   templateUrl: './login-page.page.html',
@@ -9,4 +17,19 @@ export class LoginPagePage implements OnInit {
   constructor() {}
 
   ngOnInit() {}
+
+  login() {
+    console.log('login');
+    const provider = new GoogleAuthProvider();
+    const auth = getAuth();
+    signInWithPopup(auth, provider)
+      .then((result) => {
+        const credential = GoogleAuthProvider.credentialFromResult(result);
+        const token = credential?.accessToken;
+        const user = result.user;
+      })
+      .catch((error) => {
+        console.error(error);
+      });
+  }
 }
